Move internal files to src
