+++ lib/src/model/resource_type.dart
